add rest of liquibase schema, fix bugs with models

+++ src/main/resources/db/changelog-dev/changeSet/003-create_roles_operations_table.sql
@@ -2,8 +2,8 @@
 --changeset PiotrDuma:003-create_roles_operations_table
 
 CREATE TABLE roles_operations(
-   role_id int NOT NULL,
-   operation_id int NOT NULL,
+   role_id BIGINT NOT NULL,
+   operation_id BIGINT  NOT NULL,
    FOREIGN KEY (role_id) REFERENCES roles(role_id),
    FOREIGN KEY (operation_id) REFERENCES operations(operation_id)
 );
